restructure file_based manager
